+++ src/itables/dash.py
@@ -1,21 +1,15 @@
+from itables import __version__
 from itables_for_dash import (
     ITABLE_PROPERTIES,
     ITable,
-    ITableComponent,
     ITableOutputs,
-    __version__,
-    get_itable_component_kwargs,
-    itables_for_dash_is_available,
     updated_itable_outputs,
 )
 
 __all__ = [
     "ITable",
-    "ITableComponent",
     "ITABLE_PROPERTIES",
-    "get_itable_component_kwargs",
     "ITableOutputs",
     "updated_itable_outputs",
-    "itables_for_dash_is_available",
     "__version__",
 ]