+++ tests/test_documentation_notebooks_run.py
@@ -1,4 +1,3 @@
-import sys
 from pathlib import Path
 
 import jupytext
@@ -12,8 +11,6 @@ try:
     import polars as pl
 except ImportError:
     pl = None
-
-pytestmark = pytest.mark.skipif(sys.version_info < (3, 8), reason="Require Python>=3.8")
 
 
 def list_doc_notebooks():