+++ docs/changelog.md
@@ -1,6 +1,16 @@
 ITables ChangeLog
 =================
 
+2.4.0-dev
+---------
+
+**Added**
+- We have added type hints for the `show` function and for the various app components. A SyntaxWarning is issued if either the argument name or type does not match when `warn_on_undocumented_option=True` (the default when `typeguard>=4.4.1` is installed).
+
+**Changed**
+- ITable now requires Python 3.9, due to the addition of type hints.
+
+
 2.3.0 (2025-04-05)
 ------------------